+++ CidadeInteligente.Application/Queries/GetAreaById/GetAreaByIdQueryResult.cs
@@ -1,3 +1,3 @@
 ﻿namespace CidadeInteligente.Application.Queries.GetAreaById;
 
-public record GetAreaByIdQueryResult(long AreaId, string Description);
+public record GetAreaByIdQueryResult(int AreaId, string Description);
