+++ CidadeInteligente.Application/Queries/GetProjects/GetProjectsQueryResult.cs
@@ -5,7 +5,7 @@ namespace CidadeInteligente.Application.Queries.GetProjects;
 public record GetProjectsQueryResult(int Page, int TotalPages, int ItemsCount, IEnumerable<GetProjectsQueryResult.ProjectViewModel> Data)
     : PaginationResult<GetProjectsQueryResult.ProjectViewModel>(Page, TotalPages, ItemsCount, Data)
 {
-    public record ProjectViewModel(long ProjectId, string Title, string? Description, IList<ProjectViewModel.MediaViewModel> Medias)
+    public record ProjectViewModel(int ProjectId, string Title, string? Description, IList<ProjectViewModel.MediaViewModel> Medias)
     {
         public string Description { get; private set; } = string.IsNullOrWhiteSpace(Description)
             ? string.Empty
@@ -13,7 +13,7 @@ public record GetProjectsQueryResult(int Page, int TotalPages, int ItemsCount, I
                 ? $"{Description![0..125]}..."
                 : Description;
 
-        public record MediaViewModel(long MediaId, string FileName)
+        public record MediaViewModel(int MediaId, string FileName)
         {
             public string Extension => System.IO.Path.GetExtension(FileName);
             public string Path => $"{Environment.GetEnvironmentVariable("AzureStorageBlobURL")}/{FileName}";
