+++ CidadeInteligente.Application/Queries/GetRelatedProjectsFromUser/GetRelatedProjectsFromUserQuery.cs
@@ -2,4 +2,4 @@
 
 namespace CidadeInteligente.Application.Queries.GetRelatedProjectsFromUser;
 
-public record GetRelatedProjectsFromUserQuery(long UserId, int Page) : IRequest<GetRelatedProjectsFromUserQueryResult?> { }
+public record GetRelatedProjectsFromUserQuery(int UserId, int Page) : IRequest<GetRelatedProjectsFromUserQueryResult?> { }
