+++ CidadeInteligente.Core/Repositories/IAreaRepository.cs
@@ -7,6 +7,6 @@ public interface IAreaRepository
     Task AddAsync(Area area);
     void Delete(Area area);
     Task<List<Area>> GetAllAsync();
-    Task<Area?> GetByIdAsync(long areaId, bool tracking = false);
-    Task<bool> HaveProjectsAsync(long areaId);
+    Task<Area?> GetByIdAsync(int areaId, bool tracking = false);
+    Task<bool> HaveProjectsAsync(int areaId);
 }
