+++ CidadeInteligente.Mvc/Requests/v1/UpdateUserRequest.cs
@@ -2,4 +2,4 @@
 
 namespace CidadeInteligente.Mvc.Requests.v1;
 
-public record UpdateUserRequest(long CourseId, string Name, string Email, Role Role);
+public record UpdateUserRequest(int CourseId, string Name, string Email, Role Role);
